changed IntWritable to LongWritable.

+++ src/main/java/samples/MaxValueMapper.java
@@ -2,12 +2,11 @@ package samples;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
-public class MaxValueMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
+public class MaxValueMapper extends Mapper<LongWritable, Text, Text, LongWritable> {
 
 	private static final int MISSING = 9999;
 	@Override
@@ -15,15 +14,15 @@ public class MaxValueMapper extends Mapper<LongWritable, Text, Text, IntWritable
 			throws IOException,InterruptedException{
 		String line = value.toString();
 		String year = line.substring(15,19);
-		int airTemperature;
+		long airTemperature;
 		if(line.charAt(87) == '+'){
-			airTemperature = Integer.parseInt(line.substring(88,92));
+			airTemperature = Long.parseLong(line.substring(88,92));
 		}else{
-			airTemperature= Integer.parseInt(line.substring(87, 92));
+			airTemperature= Long.parseLong(line.substring(87, 92));
 		}
 		String quality = line.substring(92, 93);
 		if(airTemperature != MISSING && quality.matches("[01459]")){
-			context.write(new Text(year), new IntWritable(airTemperature));
+			context.write(new Text(year), new LongWritable(airTemperature));
 		}
 	}
 }
