+++ src/main/java/samples/MaxValueReducer.java
@@ -2,20 +2,20 @@ package samples;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class MaxValueReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class MaxValueReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 	
 	@Override
-	public void reduce(Text key,Iterable<IntWritable> values, Context context) throws IOException, InterruptedException{
+	public void reduce(Text key,Iterable<LongWritable> values, Context context) throws IOException, InterruptedException{
 		
-		int maxValue = Integer.MIN_VALUE;
-		for(IntWritable value : values){
+		long maxValue = Integer.MIN_VALUE;
+		for(LongWritable value : values){
 			maxValue = Math.max(maxValue, value.get());
 		}
-		context.write(key, new IntWritable(maxValue));
+		context.write(key, new LongWritable(maxValue));
 		
 	}
